ai/generative_ai/requirements.txt
nats-py>=2.7.2
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
cohere==4.37
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.1